move MemBlockCache to own module outside of testing
