+++ zcash_client_memory/src/lib.rs
@@ -45,8 +45,10 @@ pub mod wallet_read;
 pub mod wallet_write;
 pub(crate) use types::*;
 
+mod block_source;
 #[cfg(test)]
 pub mod testing;
+pub use block_source::*;
 
 /// The maximum number of blocks the wallet is allowed to rewind. This is
 /// consistent with the bound in zcashd, and allows block data deeper than
